@@ -1,38 +1,49 @@
 {
-	"name": "ublaboo/api-router-project",
-	"type": "project",
-	"description": "Example starter project for ublaboo/api-router",
-	"keywords": ["rest", "api", "routes", "nette", "router", "restapi", "restrouter", "routing", "route", "starter", "project", "example"],
-	"homepage": "https://github.com/ublaboo/api-router-project",
-	"license": ["MIT"],
-	"support": {
-		"issues": "https://github.com/ublaboo/api-router-project/issues"
-	},
-	"authors": [
-		{
-			"name": "Pavel Janda",
-			"homepage": "http://paveljanda.com"
-		}
+	"name": "contributte/api-router-project",
+	"description": "Nette project example for contributte/api-router library by @f3l1x and @paveljanda",
+	"keywords": [
+		"php",
+		"nette",
+		"contributte",
+		"sandbox",
+		"project",
+		"skeleton",
+		"api"
 	],
+	"license": "MIT",
+	"type": "project",
 	"require": {
-		"php": ">= 7.1",
-		"nette/application": "^2.4",
-		"nette/bootstrap": "^2.4.2",
-		"nette/caching": "^2.5",
-		"nette/di": "^2.4",
-		"nette/http": "^2.4",
-		"nette/security": "^2.4",
-		"nette/utils": "^2.4",
-		"tracy/tracy": "^2.4",
-		"ublaboo/api-router": "~2.0.0"
+		"php": ">=7.4",
+		"contributte/application": "^0.5.0",
+		"contributte/bootstrap": "^0.5.0",
+		"contributte/cache": "^0.6.0",
+		"contributte/di": "^0.5.0",
+		"contributte/http": "^0.5.0",
+		"contributte/security": "^0.5.0",
+		"contributte/utils": "^0.5.0",
+		"contributte/tracy": "^0.5.0",
+		"contributte/api-router": "^4.1.0"
 	},
 	"require-dev": {
-		"nette/tester": "^1.6"
+		"ninjify/qa": "^0.12",
+		"ninjify/nunjuck": "^0.4",
+		"contributte/code-rules": "^1.0",
+		"contributte/dev": "^0.3",
+		"phpstan/phpstan": "^0.12.0",
+		"phpstan/phpstan-deprecation-rules": "^0.12.0",
+		"phpstan/phpstan-nette": "^0.12.0",
+		"phpstan/phpstan-strict-rules": "^0.12.0"
 	},
 	"autoload": {
 		"psr-4": {
 			"App\\": "app"
 		}
 	},
-	"minimum-stability": "stable"
+	"minimum-stability": "dev",
+	"prefer-stable": true,
+	"config": {
+		"platform": {
+			"php": "7.4"
+		}
+	}
 }
