@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Controllers;
 
-use App\Http\ApiResponse;
 use Nette\Application\IResponse;
 use Nette\Application\Request;
+use Nette\Application\Responses\JsonResponse;
 
 final class ErrorController extends AbstractController
 {
@@ -14,10 +14,6 @@ final class ErrorController extends AbstractController
 	{
 		$exception = $request->getParameter('exception');
 
-		/**
-		 * @todo Log exception
-		 */
-
-		return new ApiResponse($this->apiResponseFormatter->formatException($exception));
+		return new JsonResponse($this->apiResponseFormatter->formatException($exception));
 	}
 }
